refactor(node): decouple JWT server middleware from botas-core
Move server-specific auth middleware out of botas-core to keep it
framework-agnostic:

- botAuthExpress() moved to botas-express package
- botAuthHono() moved inline to the hono sample
- botas-core retains validateBotToken, BotAuthError, validateServiceUrl

Update docs, samples, and tests accordingly.
