+++ node/packages/botas-express/src/index.ts
@@ -3,6 +3,7 @@
 
 export { BotApp } from './bot-app.js'
 export type { BotAppOptions } from './bot-app.js'
+export { botAuthExpress } from './bot-auth-express.js'
 
 // Re-export core types so consumers only need one import
 export {
