+++ node/packages/botas/src/bot-auth-middleware.ts
@@ -241,80 +241,4 @@ export class BotAuthError extends Error {
   }
 }
 
-// ── Express middleware ────────────────────────────────────────────────────────
 
-type ExpressRequest = { headers: Record<string, string | string[] | undefined> }
-type ExpressResponse = { status(code: number): ExpressResponse; end(msg?: string): void }
-type NextFn = (err?: unknown) => void
-
-/**
- * Express middleware that validates the Bot Framework JWT token.
- *
- * @param appId - The bot's client ID. Falls back to CLIENT_ID env var.
- *
- * @example
- * server.post('/api/messages', botAuthExpress(), (req, res) => bot.processAsync(req, res))
- */
-export function botAuthExpress (
-  appId?: string
-): (req: ExpressRequest, res: ExpressResponse, next: NextFn) => Promise<void> {
-  // #95: Validate required env vars at startup, not at first request
-  const audience = appId ?? process.env['CLIENT_ID']
-  if (!audience) {
-    throw new Error('botAuthExpress: CLIENT_ID environment variable (or appId parameter) is required when auth is enabled')
-  }
-  return async (req, res, next) => {
-    const header = req.headers['authorization'] as string | undefined
-    try {
-      await validateBotToken(header, appId)
-      next()
-    } catch (err: unknown) {
-      if (err instanceof BotAuthError) {
-        res.status(401).end(err.message)
-      } else {
-        next(err)
-      }
-    }
-  }
-}
-
-// ── Hono middleware ───────────────────────────────────────────────────────────
-
-type HonoContext = {
-  req: { header(name: string): string | undefined }
-  text(body: string, status?: number): Response
-}
-type HonoNext = () => Promise<Response | void>
-
-/**
- * Hono middleware that validates the Bot Framework JWT token.
- *
- * @param appId - The bot's client ID. Falls back to CLIENT_ID env var.
- *
- * @example
- * honoApp.post('/api/messages', botAuthHono(), async (c) => {
- *   await bot.processBody(await c.req.text())
- *   return c.json({})
- * })
- */
-export function botAuthHono (
-  appId?: string
-): (c: HonoContext, next: HonoNext) => Promise<Response | void> {
-  // #95: Validate required env vars at startup, not at first request
-  const audience = appId ?? process.env['CLIENT_ID']
-  if (!audience) {
-    throw new Error('botAuthHono: CLIENT_ID environment variable (or appId parameter) is required when auth is enabled')
-  }
-  return async (c, next) => {
-    const header = c.req.header('authorization')
-    try {
-      await validateBotToken(header, appId)
-      return next()
-    } catch (err) {
-      if (err instanceof BotAuthError) {
-        return c.text(err.message, 401)
-      }
-      throw err
-    }
-  }
-}
